backend/requirements.txt
django==5.1.4
requests==2.32.3
psycopg2-binary
djangorestframework
django-extensions
django-environ
python-dateutil
django-cors-headers